Randomness modified
randomness now creates decimals and not integers

+++ src/Driver/EquationFactory.pde
@@ -12,30 +12,30 @@ class EquationFactory {
   */
   Equation randomEquation() {  
     // Random constants that are applied to the equation
-    int a = 0;
-    int b = 0;
-    int c = 0;
-    int d = 0;
-    int e = 0;
-    int f = 0;
+    float a = 0;
+    float b = 0;
+    float c = 0;
+    float d = 0;
+    float e = 0;
+    float f = 0;
     
     // Continue to generate, to eliminate boring equations
     while((a == 0 || c == 0) || (e == 0 || f == 0)) {
-      a = (int) random(-10, 10);
-      b = (int) random(-100, 100);
-      c = (int) random(-10, 10);
-      d = (int) random(-100, 100);
-      e = (int) random(-4, 5);
-      f = (int) random(-4, 5);
+      a = random(-10, 10);
+      b = random(-100, 100);
+      c = random(-10, 10);
+      d = random(-100, 100);
+      e = random(-4, 5);
+      f = random(-4, 5);
     }
     
     // Need to make variables final to use in an anonymous class
-    final int c1 = a;
-    final int c2 = b;
-    final int c3 = c;
-    final int c4 = d;
-    final int c5 = e;
-    final int c6 = f;
+    final float c1 = a;
+    final float c2 = b;
+    final float c3 = c;
+    final float c4 = d;
+    final float c5 = e;
+    final float c6 = f;
     
     
     // Create the equation with the new constants
